chore: refactor symtable a bit
