+++ server/common/src/main/java/org/eclipse/lsp/cobol/common/model/tree/variable/MultiTableDataNameNode.java
@@ -18,6 +18,8 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 import org.eclipse.lsp.cobol.common.model.Locality;
+import org.eclipse.lsp.cobol.common.model.tree.variables.IndexItemNode;
+import org.eclipse.lsp.cobol.common.utils.VariableUtils;
 
 /**
  * This value class represents the multi-dimensional Table variable that may have nested variables,
@@ -41,6 +43,24 @@ public class MultiTableDataNameNode extends VariableWithLevelNode implements Usa
     super(location, level, name, redefines, VariableType.MULTI_TABLE_DATA_NAME, global);
     this.occursClause = occursClause;
     this.usageFormat = usageFormat;
+  }
+
+  public static MultiTableDataNameNode fromDefinition(VariableDefinitionNode definitionNode) {
+    MultiTableDataNameNode variable =
+            new MultiTableDataNameNode(
+                    definitionNode.getLocality(),
+                    definitionNode.getLevel(),
+                    VariableUtils.getName(definitionNode),
+                    definitionNode.hasRedefines(),
+                    definitionNode.getOccursClauses().get(0),
+                    definitionNode.getUsage(),
+                    definitionNode.isGlobal());
+    VariableUtils.createVariableNameNode(variable, definitionNode.getVariableName());
+    for (VariableNameAndLocality nameAndLocality : definitionNode.getOccursIndexes()) {
+      variable.addChild(new IndexItemNode(
+                      nameAndLocality.getLocality(), nameAndLocality.getName(), variable.isGlobal()));
+    }
+    return variable;
   }
 
   @Override
