+++ server/common/src/main/java/org/eclipse/lsp/cobol/common/model/tree/variable/TableDataNameNode.java
@@ -19,6 +19,8 @@ import lombok.Getter;
 import lombok.ToString;
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.lsp.cobol.common.model.Locality;
+import org.eclipse.lsp.cobol.common.model.tree.variables.IndexItemNode;
+import org.eclipse.lsp.cobol.common.utils.VariableUtils;
 
 /** This value class represents the Table variable that may have an optional index */
 @Getter
@@ -91,7 +93,33 @@ public class TableDataNameNode extends ElementaryNode {
     this.occursTimes = occursTimes;
   }
 
-  @Override
+  public static TableDataNameNode fromDefinition(VariableDefinitionNode definitionNode) {
+    TableDataNameNode variable =
+            new TableDataNameNode(
+                    definitionNode.getLocality(),
+                    definitionNode.getLevel(),
+                    VariableUtils.getName(definitionNode),
+                    definitionNode.hasRedefines(),
+                    definitionNode.isGlobal(),
+                    definitionNode.getPic(),
+                    definitionNode.getValue(),
+                    definitionNode.getOccursNumber(),
+                    definitionNode.getUsage(),
+                    definitionNode.isBlankWhenZeroPresent(),
+                    definitionNode.isSignClausePresent(),
+                    definitionNode.isDynamicLength(),
+                    definitionNode.isJustified(),
+                    definitionNode.isUnBounded());
+    VariableUtils.createVariableNameNode(variable, definitionNode.getVariableName());
+    for (VariableNameAndLocality nameAndLocality : definitionNode.getOccursIndexes()) {
+      variable.addChild(new IndexItemNode(
+                      nameAndLocality.getLocality(), nameAndLocality.getName(), variable.isGlobal()));
+    }
+    return variable;
+
+  }
+
+    @Override
   protected String getVariableDisplayString() {
     StringBuilder stringBuilder = new StringBuilder(getFormattedSuffix());
     stringBuilder.append(String.format(" OCCURS %1$d TIMES", occursTimes));
