+++ server/common/src/main/java/org/eclipse/lsp/cobol/common/model/tree/variable/VariableWithLevelNode.java
@@ -19,6 +19,7 @@ import lombok.Getter;
 import lombok.ToString;
 import org.eclipse.lsp.cobol.common.model.Locality;
 import org.eclipse.lsp.cobol.common.model.tree.Node;
+import org.eclipse.lsp.cobol.common.utils.VariableUtils;
 
 /** The abstract class for variables with level number. */
 @Getter
