+++ server/common/src/main/java/org/eclipse/lsp/cobol/common/model/tree/variables/MnemonicNameNode.java
@@ -18,8 +18,10 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 import org.eclipse.lsp.cobol.common.model.Locality;
+import org.eclipse.lsp.cobol.common.model.tree.variable.VariableDefinitionNode;
 import org.eclipse.lsp.cobol.common.model.tree.variable.VariableNode;
 import org.eclipse.lsp.cobol.common.model.tree.variable.VariableType;
+import org.eclipse.lsp.cobol.common.utils.VariableUtils;
 
 /**
  * This value class represents a special name. They may substitute environment names in the
@@ -35,6 +37,15 @@ public class MnemonicNameNode extends VariableNode {
   public MnemonicNameNode(Locality location, String systemName, String name) {
     super(location, name, VariableType.MNEMONIC_NAME, false);
     this.systemName = systemName;
+  }
+
+  public static MnemonicNameNode fromDefinition(VariableDefinitionNode definitionNode) {
+    MnemonicNameNode variable = new MnemonicNameNode(
+            definitionNode.getLocality(),
+            definitionNode.getSystemName(),
+            VariableUtils.getName(definitionNode));
+    VariableUtils.createVariableNameNode(variable, definitionNode.getVariableName());
+    return variable;
   }
 
   @Override
