+++ server/common/src/main/java/org/eclipse/lsp/cobol/common/symbols/VariableAccumulator.java
@@ -15,27 +15,18 @@
 package org.eclipse.lsp.cobol.common.symbols;
 
 import org.eclipse.lsp.cobol.common.model.tree.CodeBlockDefinitionNode;
-import org.eclipse.lsp.cobol.common.model.tree.Node;
 import org.eclipse.lsp.cobol.common.model.tree.ProgramNode;
 import org.eclipse.lsp.cobol.common.model.tree.variable.VariableNode;
 
 /** Service to handle symbol information and dependencies */
 public interface VariableAccumulator {
-
-  /**
-   * * Register variable definitions into nearest ProgramNode
-   *
-   * @param node the node with VariableDefinitionNodes
-   */
-  void registerVariablesInProgram(Node node);
-
   /**
    * Add the variable definition to that program context.
    *
    * @param programNode the program where this variable belongs to.
    * @param node the variable definition node
    */
-  void addVariableDefinition(ProgramNode programNode, VariableNode node);
+  void addVariable(ProgramNode programNode, VariableNode node);
 
   /**
    * Add a paragraph defined in the program context.
