+++ server/engine/src/main/java/org/eclipse/lsp/cobol/core/engine/processors/FileDescriptionProcess.java
@@ -18,12 +18,17 @@ import org.apache.commons.lang3.StringUtils;
 import org.eclipse.lsp.cobol.common.error.ErrorSeverity;
 import org.eclipse.lsp.cobol.common.error.SyntaxError;
 import org.eclipse.lsp.cobol.common.message.MessageTemplate;
+import org.eclipse.lsp.cobol.common.model.NodeType;
+import org.eclipse.lsp.cobol.common.model.tree.Node;
+import org.eclipse.lsp.cobol.common.model.tree.variable.VariableNode;
 import org.eclipse.lsp.cobol.common.processor.ProcessingContext;
 import org.eclipse.lsp.cobol.common.processor.Processor;
+import org.eclipse.lsp.cobol.core.engine.processors.utils.SectionNodeProcessorHelper;
 import org.eclipse.lsp.cobol.core.engine.symbols.SymbolAccumulator;
 import org.eclipse.lsp.cobol.common.model.tree.variables.FileDescriptionNode;
 
 import static org.eclipse.lsp.cobol.common.VariableConstants.FD_WITHOUT_FILE_CONTROL;
+import static org.eclipse.lsp.cobol.common.model.tree.Node.hasType;
 
 /** FileDescriptionNode processor */
 public class FileDescriptionProcess implements Processor<FileDescriptionNode> {
@@ -34,14 +39,23 @@ public class FileDescriptionProcess implements Processor<FileDescriptionNode> {
   }
 
   @Override
-  public void accept(FileDescriptionNode node, ProcessingContext ctx) {
-    if (StringUtils.isBlank(node.getFileControlClause())) {
+  public void accept(FileDescriptionNode fileDescriptionNode, ProcessingContext ctx) {
+    if (StringUtils.isBlank(fileDescriptionNode.getFileControlClause())) {
       SyntaxError error =
-          node.getError(
-              MessageTemplate.of(FD_WITHOUT_FILE_CONTROL, node.getName()), ErrorSeverity.ERROR);
+          fileDescriptionNode.getError(
+              MessageTemplate.of(FD_WITHOUT_FILE_CONTROL, fileDescriptionNode.getName()), ErrorSeverity.ERROR);
       ctx.getErrors().add(error);
     }
-    ctx.getErrors().addAll(SectionNodeProcessorHelper.processNodeWithVariableDefinitions(node));
-    symbolAccumulator.registerVariablesInProgram(node);
+    ctx.getErrors().addAll(SectionNodeProcessorHelper.processNodeWithVariableDefinitions(fileDescriptionNode));
+
+    if (ctx.getCurrentProgramNode() == null) {
+      return;
+    }
+    fileDescriptionNode.getChildren().stream()
+            .flatMap(Node::getDepthFirstStream)
+            .filter(hasType(NodeType.VARIABLE))
+            .map(VariableNode.class::cast)
+            .forEach(v -> symbolAccumulator.addVariable(ctx.getCurrentProgramNode(), v));
+
   }
 }
