+++ server/engine/src/main/java/org/eclipse/lsp/cobol/core/engine/processors/implicit/ImplicitVariablesProcessor.java
@@ -47,10 +47,10 @@ public class ImplicitVariablesProcessor implements Processor<SectionNode> {
   }
 
   private void registerVariable(VariableAccumulator variableAccumulator, ProgramNode programNode, VariableNode variable) {
-    variableAccumulator.addVariableDefinition(programNode, variable);
+    variableAccumulator.addVariable(programNode, variable);
     variable.getChildren()
             .stream()
             .map(VariableNode.class::cast)
-            .forEach(c -> variableAccumulator.addVariableDefinition(programNode, c));
+            .forEach(c -> variableAccumulator.addVariable(programNode, c));
   }
 }
