+++ server/engine/src/main/java/org/eclipse/lsp/cobol/implicitDialects/cics/processor/CICSImplicitVariablesProcessor.java
@@ -52,9 +52,9 @@ public class CICSImplicitVariablesProcessor implements Processor<SectionNode> {
 
   private void registerVariable(
       VariableAccumulator variableAccumulator, ProgramNode programNode, VariableNode variable) {
-    variableAccumulator.addVariableDefinition(programNode, variable);
+    variableAccumulator.addVariable(programNode, variable);
     variable.getChildren().stream()
         .map(VariableNode.class::cast)
-        .forEach(c -> variableAccumulator.addVariableDefinition(programNode, c));
+        .forEach(c -> variableAccumulator.addVariable(programNode, c));
   }
 }
