+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/DB3014.2.cbl
@@ -4,7 +4,7 @@ DATA_NAMES
 20:10-20:14	TFIL	14:18-14:22
 
 PROCEDURES
-29:7-32:47	BUGGING-2	
+29:7-29:16	BUGGING-2	
 39:7-39:21	DB301M-CONTROL	
 38:7-38:19	DB301M-FLAGS
