+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/DB3024.2.cbl
@@ -4,7 +4,7 @@ DATA_NAMES
 22:10-22:14	TFIL	15:18-15:22
 
 PROCEDURES
-30:7-32:47	BUGGING	
+30:7-30:14	BUGGING	
 37:7-37:18	DB302-FLAGS	
 39:7-39:21	DB302M-CONTROL	32:32-32:46
