+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IC2334.2.cbl
@@ -105,7 +105,7 @@ PROCEDURES
 262:7-262:17	OPEN-FILES	
 275:7-275:11	PASS	
 279:7-279:19	PRINT-DETAIL	391:19-391:31
-199:7-259:16	SECT-IC233A-001	
+199:7-199:22	SECT-IC233A-001	
 496:7-496:24	SECT-IC233A-1-001	
 375:7-375:25	SECT-IC233A-1R-001	
 272:7-272:21	TERMINATE-CALL
