+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IC2344.2.cbl
@@ -90,11 +90,11 @@ PROCEDURES
 314:7-314:25	FAIL-ROUTINE-WRITE	307:52-307:70,308:53-308:71
 247:7-247:19	HEAD-ROUTINE	222:20-222:32
 230:7-230:12	INSPT	
-471:7-474:28	NON-GLOBAL-SECTION	
+471:7-471:25	NON-GLOBAL-SECTION	
 218:7-218:17	OPEN-FILES	
 231:7-231:11	PASS	343:27-343:31
 235:7-235:19	PRINT-DETAIL	356:19-356:31
-212:7-215:28	SECT-IC234A-001	
+212:7-212:22	SECT-IC234A-001	
 476:7-476:24	SECT-IC234A-1-001	
 331:7-331:25	SECT-IC234A-1R-001	
 595:7-595:24	SECT-IC234A-2-001	
@@ -108,7 +108,7 @@ PROCEDURES
 596:7-596:17	USE-INIT-1	
 709:7-709:17	USE-INIT-1	
 473:7-473:15	USE-PARA	
-590:7-593:29	USE-TEST	
+590:7-590:15	USE-TEST	
 340:7-340:17	USE-TEST-0	
 592:7-592:17	USE-TEST-1	
 214:7-214:17	USE-TEST-2
