+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IC4014.2.cbl
@@ -11,7 +11,7 @@ PROCEDURES
 37:7-37:20	IC401M-CANCEL	34:19-34:32
 33:7-33:21	IC401M-CONTROL	
 32:7-32:21	IC401M-NONDECL	
-27:7-28:60	IC401M-USE	
+27:7-27:17	IC401M-USE	
 
 PROGRAMS
 3	IC401M	74
