+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1034.2.cbl
@@ -170,7 +170,7 @@ PROCEDURES
 471:7-471:21	SECT-IX-03-001	
 363:7-363:21	TERMINATE-CALL	
 361:7-361:21	TERMINATE-CCVS	
-315:7-321:16	USE-IX103-TEST	
+315:7-315:21	USE-IX103-TEST	
 320:7-320:19	USE-PAR-EXIT	
 318:7-318:18	USE-PAR-001	
 424:7-424:17	WRITE-LINE	374:55-374:65,375:47-375:57,383:51-383:61,384:51-384:61,385:51-385:61,386:51-386:61,388:50-388:60,389:50-389:60,390:53-390:63,392:53-392:63,394:50-394:60,403:50-403:60,411:19-411:29,417:50-417:60,422:51-422:61,423:50-423:60,451:50-451:60,455:53-455:63,457:52-457:62,466:50-466:60
