+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1044.2.cbl
@@ -155,7 +155,7 @@ PROCEDURES
 482:7-482:25	FAIL-ROUTINE-WRITE	475:26-475:44,476:53-476:71
 410:7-410:19	HEAD-ROUTINE	366:19-366:31
 393:7-393:12	INSPT	
-331:7-349:16	IX-FS2-01	
+331:7-331:16	IX-FS2-01	
 348:7-348:21	IX-FS2-01-EXIT	338:20-338:34
 333:7-333:19	IX-FS2-01-01	
 339:7-339:19	IX-FS2-01-03	335:20-335:32
