+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1084.2.cbl
@@ -196,7 +196,7 @@ PROCEDURES
 509:7-509:25	FAIL-ROUTINE-WRITE	502:26-502:44,503:53-503:71
 437:7-437:19	HEAD-ROUTINE	393:19-393:31
 420:7-420:12	INSPT	
-358:7-376:16	IX-FS2-01	
+358:7-358:16	IX-FS2-01	
 375:7-375:21	IX-FS2-01-EXIT	365:20-365:34
 360:7-360:19	IX-FS2-01-01	
 366:7-366:19	IX-FS2-01-03	362:20-362:32
