+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1094.2.cbl
@@ -232,7 +232,7 @@ PROCEDURES
 988:7-988:24	REA-WRITE-F1-04-0	985:17-985:34
 1005:7-1005:24	REA-WRITE-F1-05-0	1002:17-1002:34
 1022:7-1022:24	REA-WRITE-F1-06-0	1019:20-1019:37
-342:7-484:24	SECT-IX109-0002	
+342:7-342:22	SECT-IX109-0002	
 637:7-637:23	SECT-IX109A-0003	
 638:7-638:19	SEQ-INIT-010	
 528:7-528:21	TERMINATE-CALL
