+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1124.2.cbl
@@ -212,7 +212,7 @@ PROCEDURES
 833:7-833:23	RWR-TEST-GF-01-0	828:42-828:58,829:42-829:58,830:42-830:58,831:42-831:58,832:42-832:58
 836:7-836:23	RWR-TEST-GF-01-1	
 849:7-849:24	RWR-WRITE-GF-01-0	846:17-846:34
-340:7-472:16	SECT-IX105-0002	
+340:7-340:22	SECT-IX105-0002	
 625:7-625:23	SECT-IX112A-0003	
 626:7-626:19	SEQ-INIT-010	
 516:7-516:21	TERMINATE-CALL
