+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1134.2.cbl
@@ -192,7 +192,7 @@ PROCEDURES
 660:7-660:24	OPN-WRITE-GF-01-0	657:17-657:34
 517:7-517:11	PASS	659:19-659:23,696:19-696:23,723:19-723:23
 521:7-521:19	PRINT-DETAIL	661:19-661:31,698:19-698:31,703:19-703:31,725:19-725:31,749:19-749:31
-331:7-470:24	SECT-IX105-0002	
+331:7-331:22	SECT-IX105-0002	
 623:7-623:23	SECT-IX113A-0003	
 624:7-624:19	SEQ-INIT-010	
 514:7-514:21	TERMINATE-CALL
