+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1154.2.cbl
@@ -188,7 +188,7 @@ PROCEDURES
 657:7-657:24	OPN-WRITE-GF-01-0	654:17-654:34
 514:7-514:11	PASS	656:19-656:23,677:19-677:23
 518:7-518:19	PRINT-DETAIL	658:19-658:31,679:19-679:31,703:19-703:31
-328:7-467:24	SECT-IX105-0002	
+328:7-328:22	SECT-IX105-0002	
 620:7-620:23	SECT-IX115A-0003	
 621:7-621:19	SEQ-INIT-010	
 511:7-511:21	TERMINATE-CALL
