+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1174.2.cbl
@@ -192,7 +192,7 @@ PROCEDURES
 685:7-685:23	RWR-TEST-GF-01-0	
 691:7-691:23	RWR-TEST-GF-01-1	
 698:7-698:24	RWR-WRITE-GF-01-0	695:20-695:37
-327:7-466:24	SECT-IX105-0002	
+327:7-327:22	SECT-IX105-0002	
 619:7-619:23	SECT-IX117A-0003	
 620:7-620:19	SEQ-INIT-010	
 510:7-510:21	TERMINATE-CALL
