+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1184.2.cbl
@@ -192,7 +192,7 @@ PROCEDURES
 699:7-699:24	OPN-WRITE-GF-02-0	696:20-696:37
 513:7-513:11	PASS	655:19-655:23,676:19-676:23
 517:7-517:19	PRINT-DETAIL	657:19-657:31,678:19-678:31,703:19-703:31
-327:7-466:24	SECT-IX105-0002	
+327:7-327:22	SECT-IX105-0002	
 619:7-619:23	SECT-IX118A-0003	
 620:7-620:19	SEQ-INIT-010	
 510:7-510:21	TERMINATE-CALL
