+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1194.2.cbl
@@ -193,7 +193,7 @@ PROCEDURES
 672:7-672:23	RWR-TEST-GF-01-0	
 678:7-678:23	RWR-TEST-GF-01-1	677:54-677:70
 691:7-691:24	RWR-WRITE-GF-01-0	688:17-688:34
-330:7-470:24	SECT-IX105-0002	
+330:7-330:22	SECT-IX105-0002	
 623:7-623:23	SECT-IX119A-0003	
 624:7-624:19	SEQ-INIT-010	
 514:7-514:21	TERMINATE-CALL
