+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1204.2.cbl
@@ -187,7 +187,7 @@ PROCEDURES
 666:7-666:23	RWR-INIT-GF-01-0	
 678:7-678:23	RWR-TEST-GF-01-1	
 686:7-686:24	RWR-WRITE-GF-01-0	683:20-683:37
-329:7-469:24	SECT-IX105-0002	
+329:7-329:22	SECT-IX105-0002	
 622:7-622:23	SECT-IX120A-0003	
 623:7-623:19	SEQ-INIT-010	
 513:7-513:21	TERMINATE-CALL
