+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX1214.2.cbl
@@ -197,7 +197,7 @@ PROCEDURES
 729:7-729:23	RWR-TEST-GF-01-0	724:42-724:58,725:42-725:58,726:42-726:58,727:42-727:58,728:42-728:58
 732:7-732:23	RWR-TEST-GF-01-1	
 745:7-745:24	RWR-WRITE-GF-01-0	742:17-742:34
-344:7-476:16	SECT-IX105-0002	
+344:7-344:22	SECT-IX105-0002	
 629:7-629:23	SECT-IX121A-0003	
 630:7-630:19	SEQ-INIT-010	
 520:7-520:21	TERMINATE-CALL
