+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX2044.2.cbl
@@ -155,7 +155,7 @@ PROCEDURES
 491:7-491:25	FAIL-ROUTINE-WRITE	484:26-484:44,485:53-485:71
 419:7-419:19	HEAD-ROUTINE	375:19-375:31
 402:7-402:12	INSPT	
-340:7-358:16	IX-FD2-01	
+340:7-340:16	IX-FD2-01	
 357:7-357:21	IX-FD2-01-EXIT	347:20-347:34
 342:7-342:19	IX-FD2-01-01	
 348:7-348:19	IX-FD2-01-03	344:20-344:32
