+++ tests/test_files/Cobol85PositiveTestsSuite/compileListing/snapshot/IX2164.2.cbl
@@ -156,7 +156,7 @@ PROCEDURES
 429:7-429:19	HEAD-ROUTINE	385:19-385:31
 412:7-412:12	INSPT	
 519:7-519:19	INX-INIT-001	
-350:7-368:16	IX-FS2-01	
+350:7-350:16	IX-FS2-01	
 367:7-367:21	IX-FS2-01-EXIT	357:20-357:34
 352:7-352:19	IX-FS2-01-01	
 358:7-358:19	IX-FS2-01-03	354:20-354:32
